Add suspending JobOperator
